# minor optimize FilterInPlace