--- a/pkg/slices/slices.go
+++ b/pkg/slices/slices.go
@@ -154,7 +154,7 @@ func BinarySearch[S ~[]E, E cmp.Ordered](x S, target E) (int, bool) {
 	return slices.BinarySearch(x, target)
 }
 
-// FilterInPlace retains all elements in []E that kee(E) returns true for.
+// FilterInPlace retains all elements in []E that keep(E) returns true for.
 // The array is *mutated in place* and returned.
 // Use Filter to avoid mutation
 func FilterInPlace[E any](s []E, keep func(E) bool) []E {