Similar with native `slices.DeleteFunc`, but `FilterInPlace` uses `keep` not `del`.


This will save 14%~30% cpu cost.
```
func BenchmarkFilterInPlace(b *testing.B) {
	ss := make([]int, 0, 100)
	for i := 0; i < 100; i++ {
		ss = append(ss, i)
	}
	b.ResetTimer()

	b.Run("keep all", func(b *testing.B) {
		for i := 0; i < b.N; i++ {
			FilterInPlace(ss, func(s int) bool {
				return true
			})
		}
	})
	b.Run("filter all", func(b *testing.B) {
		for i := 0; i < b.N; i++ {
			FilterInPlace(ss, func(s int) bool {
				return false
			})
		}
	})
}

---

goos: darwin
goarch: amd64
pkg: istio.io/istio/pkg/slices
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
                            │   old.txt   │              new.txt               │
                            │   sec/op    │   sec/op     vs base               │
FilterInPlace/keep_all-12     184.4n ± 5%   158.5n ± 3%  -14.07% (p=0.002 n=6)
FilterInPlace/filter_all-12   276.4n ± 2%   192.9n ± 2%  -30.21% (p=0.002 n=6)
geomean                       225.8n        174.9n       -22.56%

                            │   old.txt    │              new.txt               │
                            │     B/op     │    B/op     vs base                │
FilterInPlace/keep_all-12     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
FilterInPlace/filter_all-12   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
geomean                                  ²               +0.00%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                            │   old.txt    │              new.txt               │
                            │  allocs/op   │ allocs/op   vs base                │
FilterInPlace/keep_all-12     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
FilterInPlace/filter_all-12   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
geomean                                  ²               +0.00%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```